feat(core): move scope and client
