+++ packages/store/src/__tests__/hooks.test.tsx
@@ -6,7 +6,7 @@ import { afterEach, describe, expect, it, vi } from "vitest";
 import { AuiProvider } from "../utils/react-assistant-context";
 import { useAuiState } from "../useAuiState";
 import { useAuiEvent } from "../useAuiEvent";
-import { PROXIED_ASSISTANT_STATE_SYMBOL } from "@assistant-ui/core/store";
+import { PROXIED_ASSISTANT_STATE_SYMBOL } from "@assistant-ui/core/store/internal";
 
 afterEach(() => {
   vi.restoreAllMocks();
